Removed composer.lock
Composer.lock should not be in a library package since it will be
ignored when you require this package in your own project.

With a removed composer.lock the latest versions can be installed
on Travis (or oldest, or both, if you change the Travis config)
and you will be warned in advance when inclusion of this package in
other projects will lead to havoc.